updating to stay just like in the challenge

+++ src/hooks/useCart.tsx
@@ -1,4 +1,4 @@
-import { createContext, ReactNode, useContext, useEffect, useState } from 'react';
+import { createContext, ReactNode, useContext, useState } from 'react';
 import { toast } from 'react-toastify';
 import { api } from '../services/api';
 import { Product, Stock } from '../types';
@@ -22,7 +22,6 @@ interface CartContextData {
 const CartContext = createContext<CartContextData>({} as CartContextData);
 
 export function CartProvider({ children }: CartProviderProps): JSX.Element {
-  const [products, setProducts] = useState<Product[]>();
   const [cart, setCart] = useState<Product[]>(() => {
     const storagedCart = localStorage.getItem('@RocketShoes:cart')
 
@@ -34,71 +33,73 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
   });
 
 
-
   const addProduct = async (productId: number) => {
     try {
 
       let response = await api.get(`/stock/${productId}`)
-      const stock = response.data
+      const stock: Stock = response.data
 
-      response = await api.get(`/products/${productId}`)
-      const product = response.data
+      if (stock.amount < 0) {
+        toast.error('Produto não encontrado')
+        return
+      }
 
-      const existInCart = cart.find(product => product.id === productId);
+      const existInCart = cart.find(productInCart => productInCart.id === productId);
 
-      if (existInCart) {
+      if (!existInCart) {
+        response = await api.get(`/products/${productId}`)
+        let product: Product = response.data
+
+        // adicionando Amount para o product
+        product = {
+          ...product,
+          amount: 1
+        }
+
+        if (stock.amount > 0) {
+          // atualizando carrinho
+          const newCart: Product[] = [
+            ...cart,
+            { ...product },
+          ];
+
+          setCart(newCart);
+          localStorage.setItem('@RocketShoes:cart', JSON.stringify(newCart));
+          toast.success('Produto adicionado ao carrinho')
+        } else {
+          toast.error('Não há mais produtos disponíveis')
+        }
+
+      } else if (existInCart) {
         updateProductAmount({
           productId,
           amount: existInCart.amount + 1
         });
-      }
-
-      if (!existInCart) {
-
-        if (stock.amount < 1) {
-          toast.error('Não há mais produtos disponíveis.')
-        }
-
-        const newCart = [
-          ...cart,
-          { ...product, amount: 1 },
-        ];
-
-        localStorage.setItem('@RocketShoes:cart', JSON.stringify(newCart));
-
-        setCart(() => {
-          const newStoragedCart = localStorage.getItem('@RocketShoes:cart')
-
-          if (newStoragedCart) {
-            return JSON.parse(newStoragedCart);
-          }
-        });
-
-        toast.success('Produto adicionado ao carrinho.')
       } else {
-        toast.error('Produto não existe.')
+        return
       }
 
     } catch {
-      toast.error('Ocorreu um erro ao adicionar o produto ao carrinho.')
+      toast.error('Erro na adição do produto')
     }
   };
 
   const removeProduct = (productId: number) => {
     try {
+      const existProduct = cart.find(product => product.id === productId)
+
+      if (!existProduct) {
+        throw new Error
+      }
+
       const filteredCart = cart.filter(product => product.id !== productId);
 
       localStorage.setItem('@RocketShoes:cart', JSON.stringify(filteredCart));
 
-      setCart(() => {
-        const newStoragedCart = localStorage.getItem('@RocketShoes:cart')
-
-        if (newStoragedCart) {
-          return JSON.parse(newStoragedCart);
-        }
-      });
+      setCart(filteredCart);
 
       toast.success('Produto removido');
+
     } catch {
       toast.error('Erro na remoção do produto');
     }
@@ -112,21 +113,15 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
       let response = await api.get(`/stock/${productId}`)
       const stock = response.data
 
-      response = await api.get(`/products/${productId}`)
-      const product = response.data
+      const productInCart = cart.find(product => product.id === productId);
 
-      const isIncrementingAmount = product && amount > product.amount
+      const isIncrementingAmount = productInCart && amount > productInCart.amount
 
-      console.log({
-        stockAmount: stock.amount,
-        queriedAmount: amount
-      })
-
-      if (!product) {
-        throw new Error
+      if (stock.amount < amount) {
+        toast.error('Quantidade solicitada fora de estoque')
       }
 
-      if (stock.amount >= amount && amount > 1) {
+      if (stock.amount >= amount && amount > 0 && productInCart) {
         // atualizando carrinho com mais um produto
         const updatedCart = cart.map(product => {
           return product.id === productId
@@ -144,12 +139,10 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
         isIncrementingAmount
           ? message = message
           : message = 'Produto removido do carrinho.'
-      } else {
-        toast.error('Impossível atualizar o carrinho para esse valor.')
       }
 
     } catch {
-      toast.error('Erro na alteração do produto');
+      toast.error('Erro na alteração de quantidade do produto');
     }
   };
